aggiornato con la mia roba dentro la cartella feature1 con un metodo in più a connection e delle classi DAO in più, poi ho spostato i file di config in config. Modificato le entity dove stava int a double per i lat e lot. LT3

+++ src/ENTITY/Point.java
@@ -1,14 +1,14 @@
 package ENTITY;
 
 public class Point {
-    private int lat;
-    private int glon;
+    private double lat;
+    private double glon;
 
-    public int getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public int getGlon() {
+    public double getGlon() {
         return glon;
     }
 }
